Move web3 tests to be under the src directory, mirroring how we structure sdk, and applying lint and VS code symbol resolution to these files.
