+++ packages/web3/src/EntitlementCache.test.ts
@@ -2,7 +2,7 @@
  * @group main
  */
 
-import { EntitlementCache, Keyable, CacheResult } from '../src/EntitlementCache'
+import { EntitlementCache, Keyable, CacheResult } from './EntitlementCache'
 
 class Key implements Keyable {
     private readonly key: string
@@ -38,47 +38,47 @@ class StringCacheResult implements CacheResult<string> {
 
 describe('EntitlementsCacheTests', () => {
     test('caches repeat positive values', async () => {
-        const cache = new EntitlementCache<Key, Boolean>({
+        const cache = new EntitlementCache<Key, boolean>({
             positiveCacheTTLSeconds: 10,
             negativeCacheTTLSeconds: 10,
         })
 
-        const onMiss = (key: Keyable): Promise<CacheResult<boolean>> => {
+        const onMiss = (_: Keyable): Promise<CacheResult<boolean>> => {
             return Promise.resolve(new BooleanCacheResult(true))
         }
 
-        var { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
+        const { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
         expect(value).toBe(true)
         expect(cacheHit).toBe(false)
 
-        var { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
-        expect(value).toBe(true)
-        expect(cacheHit).toBe(true)
+        const { value: value2, cacheHit: cacheHit2 } = await cache.executeUsingCache(
+            new Key('key'),
+            onMiss,
+        )
+        expect(value2).toBe(true)
+        expect(cacheHit2).toBe(true)
     })
 
     test('caches repeat negative values', async () => {
-        const cache = new EntitlementCache<Key, Boolean>({
+        const cache = new EntitlementCache<Key, boolean>({
             positiveCacheTTLSeconds: 10,
             negativeCacheTTLSeconds: 10,
         })
 
-        const onMiss = (key: Keyable): Promise<CacheResult<boolean>> => {
+        const onMiss = (_: Keyable): Promise<CacheResult<boolean>> => {
             return Promise.resolve(new BooleanCacheResult(false))
         }
 
-        var { value: value, cacheHit: cacheHit } = await cache.executeUsingCache(
-            new Key('key'),
-            onMiss,
-        )
+        const { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
         expect(value).toBe(false)
         expect(cacheHit).toBe(false)
 
-        var { value: value, cacheHit: cacheHit } = await cache.executeUsingCache(
+        const { value: value2, cacheHit: cacheHit2 } = await cache.executeUsingCache(
             new Key('key'),
             onMiss,
         )
-        expect(value).toBe(false)
-        expect(cacheHit).toBe(true)
+        expect(value2).toBe(false)
+        expect(cacheHit2).toBe(true)
     })
 
     test('caches non-boolean positive values', async () => {
@@ -87,17 +87,20 @@ describe('EntitlementsCacheTests', () => {
             negativeCacheTTLSeconds: 10,
         })
 
-        const onMiss = (key: Keyable): Promise<CacheResult<string>> => {
+        const onMiss = (_: Keyable): Promise<CacheResult<string>> => {
             return Promise.resolve(new StringCacheResult('value'))
         }
 
-        var { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
+        const { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
         expect(value).toBe('value')
         expect(cacheHit).toBe(false)
 
-        var { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
-        expect(value).toBe('value')
-        expect(cacheHit).toBe(true)
+        const { value: value2, cacheHit: cacheHit2 } = await cache.executeUsingCache(
+            new Key('key'),
+            onMiss,
+        )
+        expect(value2).toBe('value')
+        expect(cacheHit2).toBe(true)
     })
 
     test('caches non-boolean falsy keys', async () => {
@@ -106,60 +109,69 @@ describe('EntitlementsCacheTests', () => {
             negativeCacheTTLSeconds: 10,
         })
 
-        const onMiss = (key: Keyable): Promise<CacheResult<string>> => {
+        const onMiss = (_: Keyable): Promise<CacheResult<string>> => {
             return Promise.resolve(new StringCacheResult(''))
         }
 
-        var { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
+        const { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
         expect(value).toBe('')
         expect(cacheHit).toBe(false)
 
-        var { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
-        expect(value).toBe('')
-        expect(cacheHit).toBe(true)
+        const { value: value2, cacheHit: cacheHit2 } = await cache.executeUsingCache(
+            new Key('key'),
+            onMiss,
+        )
+        expect(value2).toBe('')
+        expect(cacheHit2).toBe(true)
     })
 
     test('positive cache values expire after ttl', async () => {
-        const cache = new EntitlementCache<Key, Boolean>({
+        const cache = new EntitlementCache<Key, boolean>({
             positiveCacheTTLSeconds: 1,
             negativeCacheTTLSeconds: 10,
         })
 
-        const onMiss = (key: Keyable): Promise<CacheResult<boolean>> => {
+        const onMiss = (_: Keyable): Promise<CacheResult<boolean>> => {
             return Promise.resolve(new BooleanCacheResult(true))
         }
 
-        var { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
+        const { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
         expect(value).toBe(true)
         expect(cacheHit).toBe(false)
 
         // Wait 5 seconds for the positive auth cache to expire
         await new Promise((f) => setTimeout(f, 5000))
 
-        var { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
-        expect(value).toBe(true)
-        expect(cacheHit).toBe(false)
+        const { value: value2, cacheHit: cacheHit2 } = await cache.executeUsingCache(
+            new Key('key'),
+            onMiss,
+        )
+        expect(value2).toBe(true)
+        expect(cacheHit2).toBe(false)
     })
 
     test('negative cache values expire after ttl', async () => {
-        const cache = new EntitlementCache<Key, Boolean>({
+        const cache = new EntitlementCache<Key, boolean>({
             positiveCacheTTLSeconds: 10,
             negativeCacheTTLSeconds: 1,
         })
 
-        const onMiss = (key: Keyable): Promise<CacheResult<boolean>> => {
+        const onMiss = (_: Keyable): Promise<CacheResult<boolean>> => {
             return Promise.resolve(new BooleanCacheResult(false))
         }
 
-        var { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
+        const { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
         expect(value).toBe(false)
         expect(cacheHit).toBe(false)
 
         // Wait 5 seconds for the positive auth cache to expire
         await new Promise((f) => setTimeout(f, 1000))
 
-        var { value, cacheHit } = await cache.executeUsingCache(new Key('key'), onMiss)
-        expect(value).toBe(false)
-        expect(cacheHit).toBe(false)
+        const { value: value2, cacheHit: cacheHit2 } = await cache.executeUsingCache(
+            new Key('key'),
+            onMiss,
+        )
+        expect(value2).toBe(false)
+        expect(cacheHit2).toBe(false)
     })
 })
