+++ packages/web3/src/Web3Constants.test.ts
@@ -1,6 +1,6 @@
 import debug from 'debug'
-import { BASE_SEPOLIA } from '../src/Web3Constants'
-import { LocalhostWeb3Provider } from '../src/LocalhostWeb3Provider'
+import { BASE_SEPOLIA } from './Web3Constants'
+import { LocalhostWeb3Provider } from './LocalhostWeb3Provider'
 const log = debug('web3:test')
 
 describe('Web3Constants', () => {
